Reorganized queries under queries foder
