--- models/utils/createPost.js
@@ -1,20 +1,0 @@
-const Post = require("../Posts");
-
-const createPost = async (post, userId) => {
-  const date = new Date();
-  const newPost = new Post({
-    date,
-    post,
-    userId,
-  });
-
-  try {
-    await newPost.save();
-    return true;
-  } catch (error) {
-    console.log(error);
-    return false;
-  }
-};
-
-module.exports = createPost;
