--- models/utils/index.js
@@ -1,12 +1,0 @@
-const createUser = require("./createUser");
-const createPost = require("./createPost");
-const createBirthChart = require("./createBirthChart");
-const { addFriend, removeFriendFromList } = require("./addFriend");
-
-module.exports = {
-  createUser,
-  createPost,
-  createBirthChart,
-  addFriend,
-  removeFriendFromList,
-};
