+++ queries/birthChartQuery.js
@@ -1,5 +1,34 @@
 const BirthChart = require("../models/BirthCharts");
 
+const createBirthChart = async (birthChart) => {
+  const newBirthchart = new BirthChart({
+    userId: birthChart.userId,
+    sun: birthChart.sun,
+    moon: birthChart.moon,
+    mercury: birthChart.mercury,
+    venus: birthChart.venus,
+    mars: birthChart.mars,
+    jupiter: birthChart.jupiter,
+    saturn: birthChart.saturn,
+    uranus: birthChart.uranus,
+    neptune: birthChart.neptune,
+    pluto: birthChart.pluto,
+    lilith: birthChart.lilith,
+    chiron: birthChart.chiron,
+    ascendant: birthChart.ascendant,
+    midheaven: birthChart.midheaven,
+    houses: birthChart.houses,
+  });
+
+  try {
+    await newBirthchart.save();
+    return true;
+  } catch (error) {
+    console.log(error);
+    return false;
+  }
+};
+
 const findBirthChart = async (userId) => {
   try {
     const birthChart = await BirthChart.find({ userId: userId });
@@ -18,4 +47,4 @@ const deleteBirthChart = async (birthChartID) => {
   }
 };
 
-module.exports = { findBirthChart, deleteBirthChart };
+module.exports = { createBirthChart, findBirthChart, deleteBirthChart };
